requirements.txt
Flask==2.3.3
boto3==1.34.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
PyPDF2==3.0.1
